[WIP] Remove async constraint for is_any_handler_alive logic by moving it as part of endpoint
