+++ src/foldend/server/endpoints/trace_endpoint.rs
@@ -1,8 +1,11 @@
 use tokio::sync::RwLockReadGuard;
 
 use super::{handler_service_endpoint::ServiceEndpoint, trace_handler_stream::TraceHandlerStream};
-use crate::{mapping::Mapping, server::Server};
-use generated_types::TraceHandlerRequest;
+use crate::{
+    mapping::Mapping,
+    server::{endpoints::get_directory_status_endpoint::GetDirectoryStatusEndpoint, Server},
+};
+use generated_types::{GetDirectoryStatusRequest, TraceHandlerRequest};
 
 pub type Request = tonic::Request<TraceHandlerRequest>;
 pub type Response = tonic::Response<TraceHandlerStream>;
@@ -28,8 +31,16 @@ impl ServiceEndpoint<Request, Response> for TraceEndpoint<'_> {
             return Err(tonic::Status::not_found(
                 "No handler registered to filesystem to trace",
             ));
-        } else if !self.server.is_any_handler_alive().await {
-            return Err(tonic::Status::not_found("No handler is alive to trace"));
+        } else {
+            let get_dir_status_endpoint = GetDirectoryStatusEndpoint {
+                request: tonic::Request::new(GetDirectoryStatusRequest {
+                    directory_path: String::new(),
+                }),
+                mapping: self.mapping,
+            };
+            if !get_dir_status_endpoint.is_any_handler_alive() {
+                return Err(tonic::Status::not_found("No handler is alive to trace"));
+            }
         }
 
         let rx_stream = self.server.convert_trace_channel_reciever_to_stream();
