+++ src/foldend/server/mod.rs
@@ -1,13 +1,11 @@
 use std::{ops::Deref, sync::Arc};
 
 use tokio::sync::{broadcast, RwLock, RwLockReadGuard};
-use tonic::Request;
 
 use crate::config::Config;
 use crate::handler_mapping::HandlerMapping;
 use crate::mapping::Mapping;
 use endpoints::trace_handler_stream;
-use generated_types::handler_service_server::HandlerService;
 
 mod endpoints;
 mod handler_service;
@@ -47,21 +45,6 @@ impl Server {
                 yield item;
             }
         })
-    }
-
-    async fn is_any_handler_alive(&self) -> bool {
-        let response =
-            self.get_directory_status(Request::new(generated_types::GetDirectoryStatusRequest {
-                directory_path: String::new(),
-            }));
-        if let Ok(response) = response.await {
-            let response = response.into_inner();
-            return response
-                .summary_map
-                .iter()
-                .any(|(_dir, handler)| handler.is_alive);
-        }
-        false
     }
 
     fn get_handler<'a>(
